+++ packages/web-vue/components/select/style/index.ts
@@ -3,4 +3,5 @@ import '../../_components/select-view/style';
 import '../../trigger/style';
 import '../../empty/style';
 import '../../checkbox/style';
+import '../../scrollbar/style';
 import './index.less';