+++ packages/web-vue/components/tree-select/tree-select.vue
@@ -93,7 +93,6 @@ import {
   ref,
   toRefs,
   StyleValue,
-  watch,
 } from 'vue';
 import useMergeState from '../_hooks/use-merge-state';
 import { LabelValue } from './interface';
@@ -346,9 +345,7 @@ export default defineComponent({
      * @en Mount container for pop-up box
      */
     popupContainer: {
-      type: [String, Object] as PropType<
-        string | HTMLElement | null | undefined
-      >,
+      type: [String, Object] as PropType<string | HTMLElement | undefined>,
     },
     /**
      * @zh 为 value 中找不到匹配项的 key 定义节点数据