+++ packages/web-vue/components/input-number/CHANGELOG.md
@@ -2,6 +2,15 @@
 changelog: true
 ```
 
+## 2.34.0
+
+`2022-07-29`
+
+### 💎 Enhancement
+
+- Support `read-only` prop ([#1408](https://github.com/arco-design/arco-design-vue/pull/1408))
+
+
 ## 2.32.1
 
 `2022-07-01`