+++ packages/web-vue/components/list/CHANGELOG.md
@@ -2,6 +2,24 @@
 changelog: true
 ```
 
+## 2.37.2
+
+`2022-09-21`
+
+### 🐛 BugFix
+
+- Fix loading more scrollbar heights in use ([#1658](https://github.com/arco-design/arco-design-vue/pull/1658))
+
+
+## 2.37.0
+
+`2022-09-16`
+
+### 💎 Enhancement
+
+- add usage demo ([#1625](https://github.com/arco-design/arco-design-vue/pull/1625))
+
+
 ## 2.29.1
 
 `2022-06-02`